travel_app/travel_api/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.11
djangorestframework==3.15.2
python-dotenv==1.0.1
django-cors-headers==4.4.0
